+++ src/js/components/AppListStatusFilterComponent.jsx
@@ -1,6 +1,8 @@
 var React = require("react/addons");
 
 var AppStatus = require("../constants/AppStatus");
+var AppsStore = require("../stores/AppsStore");
+var AppsEvents = require("../events/AppsEvents");
 
 /* TODO extract from AppStatusComponent */
 var statusNameMapping = {
@@ -10,6 +12,13 @@ var statusNameMapping = {
   [AppStatus.DELAYED]: "Delayed",
   [AppStatus.WAITING]: "Waiting"
 };
+
+function getInitialAppStatusCount() {
+  return Object.values(AppStatus).reduce(function (memo, status) {
+    memo[status] = 0;
+    return memo;
+  }, {});
+}
 
 var AppListStatusFilterComponent = React.createClass({
   displayName: "AppListStatusFilterComponent",
@@ -24,8 +33,18 @@ var AppListStatusFilterComponent = React.createClass({
 
   getInitialState: function () {
     return {
+      appsStatusesCount: getInitialAppStatusCount(),
       selectedStatus: []
     };
+  },
+
+  componentWillMount: function () {
+    AppsStore.on(AppsEvents.CHANGE, this.onAppsChange);
+  },
+
+  componentWillUnmount: function () {
+    AppsStore.removeListener(AppsEvents.CHANGE,
+      this.onAppsChange);
   },
 
   componentDidMount: function () {
@@ -34,6 +53,18 @@ var AppListStatusFilterComponent = React.createClass({
 
   componentWillReceiveProps: function () {
     this.updateFilterStatus();
+  },
+
+  onAppsChange: function () {
+    var appsStatusesCount = getInitialAppStatusCount();
+
+    AppsStore.apps.forEach(function (app) {
+      appsStatusesCount[app.status]++;
+    });
+
+    this.setState({
+      appsStatusesCount: appsStatusesCount
+    });
   },
 
   setQueryParam: function (filterStatus) {
@@ -103,6 +134,7 @@ var AppListStatusFilterComponent = React.createClass({
 
   getStatusNodes: function () {
     var state = this.state;
+
     return Object.keys(statusNameMapping).map((key, i) => {
       let optionText = statusNameMapping[key];
 
@@ -116,7 +148,9 @@ var AppListStatusFilterComponent = React.createClass({
         <li className="checkbox" key={i}>
           <input {...checkboxProps}
             onChange={this.handleChange.bind(this, key)} />
-          <label htmlFor={`status-${key}-${i}`}>{optionText}</label>
+          <label htmlFor={`status-${key}-${i}`}>
+            {optionText} ({state.appsStatusesCount[key] || 0})
+          </label>
         </li>
       );
     });